requirements.txt
python-telegram-bot==21.3
aiohttp==3.9.5
Pillow==10.4.0
python-dotenv==1.0.1
aiofiles==23.2.1
cachetools==5.3.3
motor==3.4.0
pymongo==4.7.3
